+++ cypress/e2e/licenses/license-lifecycle.cy.js
@@ -1,8 +1,4 @@
-import {
-  KeyValue,
-} from '@folio/stripes-testing';
-
-import { HeadlineInteractor as Headline } from '../../../interactors';
+import { Headline, KeyValue } from '../../../interactors';
 
 import AppInteractor from '../../support/fragments/licenses/AppInteractor';
 import LicenseFormInteractor from '../../support/fragments/licenses/LicenseFormInteractor';