+++ cypress/support/fragments/HomeInteractor.js
@@ -1,4 +1,7 @@
-import { AppListInteractor as AppList, AppListItemInteractor as AppListItem } from '../../../interactors';
+import {
+  AppList,
+  AppListItem
+} from '../../../interactors';
 
 /* The interactor for the home page */
 export default class HomeInteractor {