+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change history for stripes-erm-testing
 
-## 1.2.0 In progress
+## 2.0.0 In progress
   * Added stripes-kint-components to dev deps
   * Tweaked Harnesses so that we can export *just* the base harness without the kint intl stuff. This means that kint-components can use the base harness as well.
   * ERM-2773 Add agreement line with internal KB resource
@@ -9,6 +9,8 @@
     * extend select interactor with selectedContent filter
   * ERM-2715 Create a license and delete a license
     * add a bunch of new interactor functions and cypress commands
+  * STRIPES-870 BREAKING upgrade react to v18
+    * ERM-2995 Upgrade stripes-erm-testing React to v18
 
 ## 1.1.0 2023-02-22
   * Initial e2e cypress tests